mmap'ed file send/receive

+++ src/main/java/org/allesoft/messenger/pure/FileAcceptRequest.java
@@ -4,5 +4,5 @@ package org.allesoft.messenger.pure;
  * Created by kabramovich on 23.11.2016.
  */
 public interface FileAcceptRequest {
-    boolean accept(FTPLayer layer, String fileName, long size);
+    boolean accept(FTPLayer layer, String fileName, int size);
 }
